Fix TODO comment "Reconsider this as a part of 'tools'"
